Renamed state model files. Added block for enrolment state
