--- src/app/services/redux/app/app-state.model.ts
@@ -1,9 +1,0 @@
-import { CurrentPassesState } from '../current-passes/current-passes-state.model';
-import { UpcomingScheduleState } from '../upcoming-schedule/upcoming-schedule-state.model';
-import { UserState } from '../user/user-state.model';
-
-export interface AppState {
-    user: UserState;
-    upcomingSchedule: UpcomingScheduleState;
-    currentPasses: CurrentPassesState;
-};
