+++ src/app/services/redux/blocks-for-enrolment/blocks-for-enrolment.model.ts
@@ -1,0 +1,15 @@
+export interface BlocksForEnrolmentState {
+    isLoading: boolean;
+    errors: string[];
+
+    blocks: BlockForEnrolment[];
+}
+
+export interface BlockForEnrolment {
+    id: number;
+    name: string;
+    startDate: Date;
+    minutesPerClass: number;
+    spacesAvailable: number;
+    isAlreadyRegistered: boolean;
+}
