+++ src/app/services/redux/user/user.actions.ts
@@ -1,4 +1,4 @@
-import { userRole } from './user-state.model';
+import { userRole } from './user.model';
 import { Action } from '@ngrx/store';
 
 export const CHECK_LOGIN_STATUS = '[User] Check Login Status';
